pyproject.toml
[project]
name = "mcp-spark-events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[tool.ruff]
target-version = "py312"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
